Inline run_log_review helpers into log_audit slot; delete run_log_review
Closes deferred item #2 from issue #1292.

The legacy reflections.auditing.run_log_review callable was the entry
point for the daily-log-review registry entry. PR #1281 made the new
log_audit slot wrap its three private helpers
(`_collect_sentry_counts`, `_read_log_tail_lines`,
`_read_log_text_bounded`), leaving the old function as the parallel-run
delivery path. This commit:

- Inlines the three helpers and the size constants into
  reflections/pm_audio_briefing/log_audit.py.
- Deletes run_log_review, _send_log_review_telegram, and the
  inlined helpers from reflections/auditing.py.
- Leaves extract_structured_errors in reflections.utils because
  run_hooks_audit still imports it from there (per issue Non-Goals:
  'Other auditing.py helpers stay').

Tests:
- tests/unit/test_reflections_auditing.py — RENAMED+REWRITTEN to
  tests/unit/reflections/test_log_audit_sentry.py. The
  _send_log_review_telegram and run_log_review end-to-end tests
  were deleted (the functions are gone). The Sentry helper tests and the
  bounded-read regression test now patch the slot module's namespace.
- tests/unit/test_reflections_package.py — removed the
  run_log_review smoke test and the _read_log_text_bounded cap
  test (covered in the new test_log_audit_sentry.py); the event-loop-safe
  canary now only guards run_hooks_audit.

The legacy registry entry daily-log-review (callable
reflections.auditing.run_log_review) must be flipped to
enabled: false in the vault yaml as part of the cutover (deferred
item #4); see PR body for the operator step.
